Update to fix bug
We're seeing a bug in android only (tested in 4.4.2) where you're stopped from typing a message in the text area. The letters don't appear in the area (although symbols, periods, other non-alpha characters will show up).  Sometimes if I select a type ahead word it will appear and then more text appears with a delay. It isn't possible to type a message with this current behavior.